Bugfix for ArmController

+++ RED/ViewModels/ControlCenter/ArmControllerMode.cs
@@ -36,7 +36,7 @@ namespace RED.ViewModels.ControlCenter
             }
         }
 
-        public enum ArmAction : int
+        public enum ArmAction : byte
         {
             WristCounterclockwise = 0,
             WristClockwise = 1,
@@ -51,7 +51,7 @@ namespace RED.ViewModels.ControlCenter
             BaseServoCounterclockwise = 10,
             BaseServoClockwise = 11,
             Reset = 12,
-            Idle = Int32.MaxValue
+            Idle = Byte.MaxValue
         }
 
         public ArmControllerMode(InputViewModel inputVM, ControlCenterViewModel cc)
@@ -76,71 +76,71 @@ namespace RED.ViewModels.ControlCenter
             if (InputVM.ButtonY)
             {
                 CurrentAction = ArmAction.Reset;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
                 _controlCenter.Console.WriteToConsole("Robotic Arm Resetting...");
             }
             if (InputVM.JoyStick2X < 0)
             {
                 CurrentAction = ArmAction.WristCounterclockwise;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
             else if (InputVM.JoyStick2X > 0)
             {
                 CurrentAction = ArmAction.WristClockwise;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
             else if (InputVM.JoyStick2Y < 0)
             {
                 CurrentAction = ArmAction.WristDown;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
             else if (InputVM.JoyStick2Y > 0)
             {
                 CurrentAction = ArmAction.WristUp;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
 
             if (InputVM.JoyStick1X < 0)
             {
                 CurrentAction = ArmAction.ElbowCounterclockwise;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
             else if (InputVM.JoyStick1X > 0)
             {
                 CurrentAction = ArmAction.ElbowClockwise;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
             else if (InputVM.JoyStick1Y < 0)
             {
                 CurrentAction = ArmAction.ElbowDown;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
             else if (InputVM.JoyStick1Y > 0)
             {
                 CurrentAction = ArmAction.ElbowUp;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
 
             if (InputVM.DPadL)
             {
                 CurrentAction = ArmAction.BaseServoCounterclockwise;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
             else if (InputVM.DPadR)
             {
                 CurrentAction = ArmAction.BaseServoClockwise;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
 
             if (InputVM.DPadU)
             {
                 CurrentAction = ArmAction.BaseActuatorForward;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
             else if (InputVM.DPadD)
             {
                 CurrentAction = ArmAction.BaseActuatorBackward;
-                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), CurrentAction);
+                _controlCenter.DataRouter.Send(_controlCenter.MetadataManager.GetId("Arm"), (byte)CurrentAction);
             }
         }
